+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- hnefatafl-server-full: don't send display_users or display_games unless
+  something has changed or you log in
 - hnefatafl-client: clear the texts before logging in.
 
 ## [4.5.0] - 2025-12-18