+++ CHANGELOG.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- hnefatafl-server-full: do display_games, display_users every quarter second.
 - hnefatafl-server-full: don't send display_users or display_games unless
-  something has changed or you log in
+  something has changed or you log in.
 - hnefatafl-client: clear the texts before logging in.
 
 ## [4.5.0] - 2025-12-18