+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Use more generic logging
+- Log only the module being used by default.
+
 ### Fixed
 
 - hnefatafl-server-full: don't send an empty message.