+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - hnefatafl-client: an error message for too many connections or server quit.
-- admins to the server and client ron settings file.
+- admins to the server ron settings file.
 - Document the feature flags.
 - Make `--debug` an argument to hnefatafl-{ai, client, server-full}.
 - hnefatafl-client: optionally on windows display console out by passing