+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- hnefatafl-server-full: `--secure` flag.
 - admins to the server ron settings file.
 - Document the feature flags.
 - Make `--debug` an argument to hnefatafl-{ai, client, server-full}.